pyproject.toml
[tool.poetry]
name = "flowshiki"
version = "0.0.1a"
description = "Flow Launcher plugin to search through Shikimori"
authors = ["NoPlagiarism <37241775+NoPlagiarism@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyflowlauncher = "^0.9.2"
httpx = "^0.27.0"
httpx-auth = "^0.22.0"
loguru = "^0.7.2"
typing-extensions = "^4.12.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
